+++ src/test/kotlin/com/gemini/krakenbot/config/SettingsTest.kt
@@ -1,10 +1,13 @@
 package com.gemini.krakenbot.config
 
+import io.kotest.core.spec.IsolationMode
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.booleans.shouldBeTrue
 import io.kotest.matchers.shouldBe
 
 class SettingsTest : StringSpec() {
+    override fun isolationMode() = IsolationMode.InstancePerTest
+
     init {
         "constructor_defaultsNullValues" {
             val settings =